Rename ConfigDTO to LegacyConfigDTO
